+++ parsers/registry/parser/parser.cpp
@@ -509,7 +509,7 @@ void RegistryParser::processValueData(libregf_value_t* value_handle,
 
       while (current_pos < end_pos) {
         const size_t string_length =
-            strnlen(current_pos, end_pos - current_pos);
+            strnlen(current_pos, static_cast<size_t>(end_pos - current_pos));
         if (string_length == 0) break;
 
         strings.emplace_back(current_pos, string_length);
